feat(typingtext.jsx): implement the functionality of  TypingText component
